move NormalizeAddr from configutil to parseutil
